streamlit_backend/requirements.txt
streamlit>=1.30.0
supabase>=2.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
httpx>=0.26.0
fastapi>=0.109.0
uvicorn>=0.27.0
